feat: services folder with fetch data function
